Refactoring notifications to chart
Refactoring notifications to chart
