+++ BmiCalc/app/src/main/java/pamo/bmicalc/ui/chart/ChartFragment.java
@@ -1,4 +1,4 @@
-package pamo.bmicalc.ui.notifications;
+package pamo.bmicalc.ui.chart;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -14,17 +14,17 @@ import androidx.lifecycle.ViewModelProvider;
 
 import pamo.bmicalc.R;
 
-public class NotificationsFragment extends Fragment {
+public class ChartFragment extends Fragment {
 
-    private NotificationsViewModel notificationsViewModel;
+    private ChartViewModel chartViewModel;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        notificationsViewModel =
-                new ViewModelProvider(this).get(NotificationsViewModel.class);
-        View root = inflater.inflate(R.layout.fragment_notifications, container, false);
+        chartViewModel =
+                new ViewModelProvider(this).get(ChartViewModel.class);
+        View root = inflater.inflate(R.layout.fragment_chart, container, false);
         final TextView textView = root.findViewById(R.id.text_notifications);
-        notificationsViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
+        chartViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
             @Override
             public void onChanged(@Nullable String s) {
                 textView.setText(s);
