+++ BmiCalc/app/src/main/java/pamo/bmicalc/ui/chart/ChartViewModel.java
@@ -1,16 +1,16 @@
-package pamo.bmicalc.ui.notifications;
+package pamo.bmicalc.ui.chart;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class ChartViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public ChartViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("This is chart fragment");
     }
 
     public LiveData<String> getText() {
