+++ BmiCalc/app/src/main/res/navigation/mobile_navigation.xml
@@ -18,10 +18,10 @@
         tools:layout="@layout/fragment_quiz" />
 
     <fragment
-        android:id="@+id/navigation_notifications"
-        android:name="pamo.bmicalc.ui.notifications.NotificationsFragment"
-        android:label="@string/notifications_title"
-        tools:layout="@layout/fragment_notifications" />
+        android:id="@+id/navigation_chart"
+        android:name="pamo.bmicalc.ui.chart.ChartFragment"
+        android:label="@string/chart_title"
+        tools:layout="@layout/fragment_chart" />
     <fragment
         android:id="@+id/navigation_calculate"
         android:name="pamo.bmicalc.ui.calculate.CalculateFragment"
