+++ BmiCalc/app/src/main/res/values/strings.xml
@@ -4,7 +4,7 @@
     <!-- NAV STRINGS -->
     <string name="bmiCalc_title">Calculate</string>
     <string name="quiz_title">Quiz</string>
-    <string name="notifications_title">Notifications</string>
+    <string name="chart_title">Chart</string>
 
     <!-- HOME SCREEN ACTIVITY -->
     <string name="bmiCalc_abbr_text">BMI</string>
